app/requirements.txt
python-telegram-bot==20.1
aiosqlite==0.18.0
Jinja2==3.1.2
python-dotenv==0.21.1
pytz==2022.7.1
openpyxl==3.1.1
black==23.1.0
poetry==1.4.0